Bugfix: Metadata Tilt wird jetzt mit Double gesetzt statt int

+++ app/src/main/java/de/th/ro/datavis/models/MetaData.java
@@ -64,7 +64,7 @@ public class MetaData {
         return tilt;
     }
 
-    public void setTilt(int tilt) {
+    public void setTilt(double tilt) {
         this.tilt = tilt;
     }
